# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inter Finder API",
    "version": "1.0.0",
    "description": "API Documentation for the Inter Finder web application. This provides a clear contract for frontend and backend teams."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/interviews/company": {
      "get": {
        "summary": "Get all interviews for company's applications",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          }
        }
      }
    },
    "/interviews/intern": {
      "get": {
        "summary": "Get all interviews for the logged-in intern",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (intern only)"
          }
        }
      }
    },
...
  },
...
}